Fix path to remote scripts
Make the path to remote scripts to be the same for both the Aquilon host
and the Kayobe setup host.

And, in both cases, to be the same path where those scripts are located in this repository.
This way, the easiest way to deploy them, and to keep them up to date,
is just to clone this repository on those hosts.

README file updated accordingly.